#250 refactored timemachine module - unit tests still failing ;-(
