+++ fullstop-testing/it/fullstop/pom.xml
@@ -41,7 +41,12 @@
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>fullstop-timemachine</artifactId>
-            <version>${fullstop-timemachine.version}</version>
+            <version>${fullstop-web.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>fullstop-violations-rest</artifactId>
+            <version>${fullstop-web.version}</version>
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
