+++ fullstop-testing/it/pom.xml
@@ -32,7 +32,7 @@
         <fullstop-core.version>1.2.0-SNAPSHOT</fullstop-core.version>
         <fullstop-plugin-api.version>1.2.0-SNAPSHOT</fullstop-plugin-api.version>
         <fullstop-violation-api.version>1.2.0-SNAPSHOT</fullstop-violation-api.version>
-        <fullstop-timemachine.version>1.2.0-SNAPSHOT</fullstop-timemachine.version>
+        <fullstop-web.version>1.2.0-SNAPSHOT</fullstop-web.version>
         <fullstop-violation.version>1.2.0-SNAPSHOT</fullstop-violation.version>
         <fullstop-plugin.version>1.2.0-SNAPSHOT</fullstop-plugin.version>
     </properties>
