+++ fullstop-web/fullstop-timemachine/src/main/java/org/zalando/stups/fullstop/filter/CORSFilter.java
@@ -34,7 +34,7 @@ public class CORSFilter implements Filter {
             throws IOException, ServletException {
         HttpServletResponse response = (HttpServletResponse) res;
         response.setHeader("Access-Control-Allow-Origin", "*");
-        response.setHeader("Access-Control-Allow-Methods", "POST, GET, OPTIONS, DELETE");
+        response.setHeader("Access-Control-Allow-Methods", "POST, GET, OPTIONS, DELETE, PUT");
         response.setHeader("Access-Control-Max-Age", "3600");
         response.setHeader("Access-Control-Allow-Headers", "Content-Type, x-requested-with, Authorization");
         chain.doFilter(req, res);
