+++ fullstop-web/fullstop-timemachine/src/test/java/org/zalando/stups/fullstop/swagger/api/FullstopApiTest.java
@@ -23,46 +23,29 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.security.authentication.TestingAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.setup.StandaloneMockMvcBuilder;
-import org.zalando.stups.fullstop.common.RestControllerTestSupport;
 import org.zalando.stups.fullstop.s3.S3Service;
 import org.zalando.stups.fullstop.swagger.model.LogObj;
-import org.zalando.stups.fullstop.swagger.model.Violation;
 import org.zalando.stups.fullstop.teams.TeamOperations;
-import org.zalando.stups.fullstop.teams.Account;
 import org.zalando.stups.fullstop.violation.entity.LifecycleEntity;
-import org.zalando.stups.fullstop.violation.entity.ViolationEntity;
 import org.zalando.stups.fullstop.violation.service.ApplicationLifecycleService;
 import org.zalando.stups.fullstop.violation.service.ViolationService;
+import org.zalando.stups.fullstop.web.test.RestControllerTestSupport;
 import sun.misc.BASE64Encoder;
 
 import java.util.Date;
-import java.util.List;
-import java.util.UUID;
 
-import static com.google.common.collect.Lists.newArrayList;
 import static org.joda.time.DateTimeZone.UTC;
 import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyLong;
-import static org.mockito.Matchers.anyString;
-import static org.mockito.Mockito.eq;
-import static org.mockito.Mockito.isNull;
 import static org.mockito.Mockito.*;
-import static org.springframework.data.domain.Sort.Direction.ASC;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import static org.zalando.stups.fullstop.builder.domain.ViolationEntityBuilder.violation;
-import static org.zalando.stups.fullstop.common.test.mvc.matcher.MatcherHelper.hasSize;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.zalando.stups.fullstop.s3.LogType.USER_DATA;
 
 /**
@@ -73,8 +56,6 @@ import static org.zalando.stups.fullstop.s3.LogType.USER_DATA;
 public class FullstopApiTest extends RestControllerTestSupport {
 
     public static final String ACCOUNT_ID = "123";
-
-    public static final String COMMENT = "my comment";
 
     public static final String ENCODED_LOG_FILE = new BASE64Encoder().encode("this is my log".getBytes());
 
@@ -88,30 +69,13 @@ public class FullstopApiTest extends RestControllerTestSupport {
     private FullstopApi fullstopApiController;
 
     @Autowired
-    private ViolationService violationServiceMock;
-
-    @Autowired
-    private TeamOperations mockTeamOperations;
-
-    @Autowired
     private ApplicationLifecycleService mockApplicationLifecycleService;
-
-    private Violation violationRequest;
 
     private LogObj logObjRequest;
 
-    private ViolationEntity violationResult;
-
     @Before
     public void setUp() throws Exception {
-        reset(violationServiceMock, mockTeamOperations, mockApplicationLifecycleService);
-
-        violationRequest = new Violation();
-        violationRequest.setAccountId(ACCOUNT_ID);
-        violationRequest.setRegion(REGION);
-        violationRequest.setEventId(UUID.randomUUID().toString());
-
-        violationResult = testDataInitializer.create(violation().id(0L).version(0L));
+        reset(mockApplicationLifecycleService);
 
         logObjRequest = new LogObj();
         logObjRequest.setAccountId(ACCOUNT_ID);
@@ -133,12 +97,11 @@ public class FullstopApiTest extends RestControllerTestSupport {
     @After
     public void tearDown() throws Exception {
         SecurityContextHolder.clearContext();
-        verifyNoMoreInteractions(violationServiceMock, mockTeamOperations, mockApplicationLifecycleService);
+        verifyNoMoreInteractions(mockApplicationLifecycleService);
     }
 
     @Test
     public void testInstanceLogs() throws Exception {
-
         when(
                 mockApplicationLifecycleService.saveInstanceLogLifecycle(
                         any(),
@@ -160,143 +123,6 @@ public class FullstopApiTest extends RestControllerTestSupport {
     @Test
     public void testInstanceLogsNotBase64LogDataEncoded() throws Exception {
         // test with not encoded log data
-    }
-
-    @Test
-    public void testGetOneViolation() throws Exception {
-        violationResult.setId(1L);
-        when(violationServiceMock.findOne(1L)).thenReturn(violationResult);
-
-        final ResultActions resultActions = this.mockMvc.perform(get("/api/violations/1")).andExpect(status().isOk());
-        resultActions.andExpect(jsonPath("$.id").value(1));
-        verify(violationServiceMock).findOne(1L);
-    }
-
-    @Test
-    public void testGetOneNullViolation() throws Exception {
-        when(violationServiceMock.findOne(948439L)).thenReturn(null);
-
-        final ResultActions resultActions = this.mockMvc.perform(get("/api/violations/948439"))
-                                                        .andExpect(status().isNotFound());
-        resultActions.andExpect(content().string("\"Violation with id: 948439 not found!\""));
-        verify(violationServiceMock).findOne(948439L);
-    }
-
-    @Test
-    public void testViolations() throws Exception {
-        when(violationServiceMock.queryViolations(any(), any(), any(), any(), any(), any(), any(), any())).thenReturn(
-                new PageImpl<>(
-                        newArrayList(violationResult), new PageRequest(0, 20, ASC, "id"), 50));
-
-        final ResultActions resultActions = this.mockMvc.perform(get("/api/violations")).andExpect(status().isOk());
-
-        resultActions.andExpect(jsonPath("$.content").value(hasSize(1)));
-
-        verify(violationServiceMock).queryViolations(
-                isNull(List.class),
-                isNull(DateTime.class),
-                isNull(Long.class),
-                isNull(Boolean.class),
-                isNull(Integer.class),
-                isNull(Boolean.class),
-                isNull(String.class),
-                any());
-    }
-
-    @Test
-    public void testViolationsWithParams() throws Exception {
-
-        DateTime dateTime = new DateTime(UTC);
-        long lastViolation = 0L;
-
-        when(
-                violationServiceMock.queryViolations(
-                        eq(newArrayList("123")),
-                        any(DateTime.class),
-                        eq(lastViolation),
-                        eq(true),
-                        any(),
-                        any(),
-                        any(),
-                        any()))
-                .thenReturn(new PageImpl<>(newArrayList(violationResult), new PageRequest(0, 20, ASC, "id"), 50));
-
-        ResultActions resultActions = this.mockMvc.perform(
-                get("/api/violations?accounts=123&checked=true&last-violation=0&since=" + dateTime))
-                                                  .andExpect(status().isOk());
-
-        resultActions.andExpect(jsonPath("$.content").value(hasSize(1)));
-
-        verify(violationServiceMock).queryViolations(
-                eq(newArrayList("123")), any(DateTime.class), eq(lastViolation), eq(
-                        true), any(), any(), any(), any());
-    }
-
-    @Test
-    public void testResolveViolation() throws Exception {
-        when(violationServiceMock.findOne(anyLong())).thenReturn(violationResult);
-        when(violationServiceMock.save(eq(violationResult))).thenReturn(violationResult);
-        when(mockTeamOperations.getTeamsByUser(anyString())).thenReturn(
-                newArrayList(
-                        new Account(
-                                violationResult.getAccountId(),
-                                "Foo",
-                                "aws",
-                                "account desc")));
-
-        violationRequest.setComment("my comment");
-
-        String message = "test";
-
-        byte[] bytes = objectMapper.writeValueAsBytes(message);
-
-        this.mockMvc.perform(
-                post("/api/violations/156/resolution").contentType(APPLICATION_JSON).content(bytes))
-                    .andExpect(status().isOk());
-
-        verify(violationServiceMock).findOne(eq(156L));
-        verify(violationServiceMock).save(eq(violationResult));
-        verify(mockTeamOperations).getTeamsByUser(eq("test-user"));
-    }
-
-    @Test
-    public void testResolveUnknownViolation() throws Exception {
-        when(violationServiceMock.findOne(any(Long.class))).thenReturn(null);
-
-        String message = "test";
-
-        byte[] bytes = objectMapper.writeValueAsBytes(message);
-
-        this.mockMvc.perform(
-                post("/api/violations/156/resolution").contentType(APPLICATION_JSON).content(bytes))
-                    .andExpect(status().isNotFound());
-
-        verify(violationServiceMock).findOne(eq(156L));
-    }
-
-    @Test
-    public void testResolveOtherTeamsViolation() throws Exception {
-        when(violationServiceMock.findOne(anyLong())).thenReturn(violationResult);
-        when(mockTeamOperations.getTeamsByUser(anyString())).thenReturn(
-                newArrayList(
-                        new Account(
-                                "foo",
-                                "Foo",
-                                "other_teams_account",
-                                "aws")));
-
-        violationRequest.setComment("my comment");
-
-        String message = "test";
-
-        byte[] bytes = objectMapper.writeValueAsBytes(message);
-
-        this.mockMvc.perform(
-                post("/api/violations/156/resolution").contentType(APPLICATION_JSON).content(bytes))
-                    .andExpect(status().isForbidden());
-
-        verify(violationServiceMock).findOne(eq(156L));
-        verify(mockTeamOperations).getTeamsByUser(eq("test-user"));
     }
 
     @Override
