+++ fullstop-web/fullstop-web-test-support/src/main/java/org/zalando/stups/fullstop/web/test/TestDataInitializer.java
@@ -1,0 +1,29 @@
+/**
+ * Copyright (C) 2015 Zalando SE (http://tech.zalando.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.zalando.stups.fullstop.web.test;
+
+/**
+ * @author ahartmann
+ */
+public class TestDataInitializer {
+
+    public static final TestDataInitializer INITIALIZER = new TestDataInitializer();
+
+    public <T> T create(final TestObjectBuilder<T> entityBuilder) {
+        return entityBuilder.build();
+    }
+
+}
